Refactor: SignUpInput, postApi, postSignUp
